backend/requirements.txt
fastapi==0.115.12
pydantic==2.10.6
pydantic[email]
uvicorn==0.33.0
passlib~=1.7.4
sqlmodel
mysql-connector-python
passlib
httpx
pytest-asyncio
pyjwt
fastapi-pagination
google-genai
dotenv
tensorflow==2.19.0
tensorflow-decision-forests==1.12.0
scikit-learn==1.6.1